+++ Src/FluentAssertions/Primitives/DateTimeAssertions.cs
@@ -22,6 +22,8 @@ namespace FluentAssertions.Primitives
         }
     }
 
+#pragma warning disable CS0659 // Ignore not overriding Object.GetHashCode()
+#pragma warning disable CA1065 // Ignore throwing NotSupportedException from Equals
     /// <summary>
     /// Contains a number of methods to assert that a <see cref="DateTime"/> is in the expected state.
     /// </summary>
@@ -920,5 +922,9 @@ namespace FluentAssertions.Primitives
 
             return new AndConstraint<TAssertions>((TAssertions)this);
         }
+
+        /// <inheritdoc/>
+        public override bool Equals(object obj) =>
+            throw new NotSupportedException("Calling Equals on Assertion classes is not supported.");
     }
 }