+++ Src/FluentAssertions/Types/MethodInfoSelectorAssertions.cs
@@ -9,6 +9,8 @@ using FluentAssertions.Execution;
 
 namespace FluentAssertions.Types
 {
+#pragma warning disable CS0659 // Ignore not overriding Object.GetHashCode()
+#pragma warning disable CA1065 // Ignore throwing NotSupportedException from Equals
     /// <summary>
     /// Contains assertions for the <see cref="MethodInfo"/> objects returned by the parent <see cref="MethodInfoSelector"/>.
     /// </summary>
@@ -325,5 +327,9 @@ namespace FluentAssertions.Types
 #pragma warning disable CA1822 // Do not change signature of a public member
         protected string Context => "method";
 #pragma warning restore CA1822
+
+        /// <inheritdoc/>
+        public override bool Equals(object obj) =>
+            throw new NotSupportedException("Calling Equals on Assertion classes is not supported.");
     }
 }